重构TopBar组件，减少重复代码
