+++ glancy-site/src/components/DesktopTopBar.jsx
@@ -1,10 +1,6 @@
-import { useState, useRef, useEffect } from 'react'
-import ModelSelector from './Toolbar/ModelSelector.jsx'
 import './DesktopTopBar.css'
 import './Header/Header.css'
-import { useLanguage } from '../LanguageContext.jsx'
-import { useUserStore } from '../store/userStore.js'
-import { Link } from 'react-router-dom'
+import TopBarActions from './TopBarActions.jsx'
 
 function DesktopTopBar({
   term = '',
@@ -14,22 +10,6 @@ function DesktopTopBar({
   onToggleFavorite,
   canFavorite = false
 }) {
-  const [open, setOpen] = useState(false)
-  const menuRef = useRef(null)
-  const { t } = useLanguage()
-  const user = useUserStore((s) => s.user)
-
-  useEffect(() => {
-    function handlePointerDown(e) {
-      if (menuRef.current && !menuRef.current.contains(e.target)) {
-        setOpen(false)
-      }
-    }
-    if (open) {
-      document.addEventListener('pointerdown', handlePointerDown)
-    }
-    return () => document.removeEventListener('pointerdown', handlePointerDown)
-  }, [open])
 
   return (
     <header className="desktop-topbar">
@@ -45,41 +25,11 @@ function DesktopTopBar({
         ←
       </button>
       <div className="term-text">{term}</div>
-      <div className="topbar-right">
-        {canFavorite && (
-          <button
-            type="button"
-            className="favorite-toggle"
-            onClick={onToggleFavorite}
-          >
-            {favorited ? '★' : '☆'}
-          </button>
-        )}
-        {user ? (
-          <>
-            <ModelSelector />
-            <div className="more-menu" ref={menuRef}>
-              <button
-                type="button"
-                className="more-btn"
-                onClick={() => setOpen(!open)}
-              >
-                ⋮
-              </button>
-              {open && (
-                <div className="menu">
-                  <button type="button">
-                    <span className="icon">🔗</span>{t.share}
-                  </button>
-                  <button type="button">
-                    <span className="icon">🚩</span>{t.report}
-                  </button>
-                </div>
-              )}
-            </div>
-          </>
-        ) : null}
-      </div>
+      <TopBarActions
+        favorited={favorited}
+        onToggleFavorite={onToggleFavorite}
+        canFavorite={canFavorite}
+      />
     </header>
   )
 }
